All requested changes have been made and are working, all that there now is testing and adding the local source for the map.
