+++ package.json
@@ -16,8 +16,6 @@
     "@maptiler/sdk": "^2.4.1",
     "@tresjs/cientos": "^4.0.3",
     "@tresjs/core": "^4.3.1",
-    "@types/geojson": "^7946.0.14",
-    "@types/leaflet": "^1.9.14",
     "leaflet": "^1.9.4",
     "pinia": "^2.2.5",
     "roslib": "^1.4.1",
@@ -28,6 +26,8 @@
   },
   "devDependencies": {
     "@eslint/compat": "^1.2.2",
+    "@types/geojson": "^7946.0.16",
+    "@types/leaflet": "^1.9.16",
     "@types/roslib": "^1.3.5",
     "@types/three": "^0.170.0",
     "@vitejs/plugin-vue": "^4.6.2",
